serving/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-aiplatform==1.46.0
google-cloud-firestore==2.16.0